XrayAgent/requirements.txt
openai>=1.0.0
Pillow>=9.0.0
numpy>=1.20.0
pandas>=1.3.0
requests>=2.25.0
python-dotenv>=0.19.0

# Core ML dependencies for tool execution
torch>=1.9.0
torchvision>=0.10.0  
transformers>=4.20.0
scikit-image>=0.18.0
opencv-python>=4.5.0

# Medical imaging dependencies
pydicom>=2.2.0
torchxrayvision>=1.0.0

# Additional dependencies for notebooks
jupyter>=1.0.0
ipywidgets>=7.6.0
matplotlib>=3.3.0
seaborn>=0.11.0
tqdm>=4.60.0

# Development dependencies (optional)
pytest>=6.0.0
black>=21.0.0
flake8>=3.8.0 